restructure folder structure to include a src directory
