+++ news-aggregation/src/handlers/ExecuteBlockActionHandler.ts
@@ -2,7 +2,7 @@ import {
 	IUIKitResponse,
 	UIKitBlockInteractionContext,
 } from '@rocket.chat/apps-engine/definition/uikit';
-import { NewsAggregationApp } from '../NewsAggregationApp';
+import { NewsAggregationApp } from '../../NewsAggregationApp';
 import {
 	IHttp,
 	IModify,
@@ -10,7 +10,7 @@ import {
 	IRead,
 } from '@rocket.chat/apps-engine/definition/accessors';
 import { SubscriptionPersistence } from '../persistence/SubscriptionPersistence';
-import { ModalEnum } from '../enums/modalEnum';
+import { ModalEnum } from '../../enums/modalEnum';
 
 export class ExecuteBlockActionHandler {
 	private context: UIKitBlockInteractionContext;
