+++ news-aggregation/src/persistence/RoomPersistence.ts
