+++ news-aggregation/src/persistence/UserPersistence.ts
