+++ news-aggregation/src/processors/DailyNewsProcessor.ts
