+++ news-aggregation/src/ui-kit/block/IActionBlock.ts
