+++ news-aggregation/src/ui-kit/block/IBlockBuilder.ts
