+++ news-aggregation/src/ui-kit/block/IContextBlock.ts
