+++ news-aggregation/src/ui-kit/block/IInputBlock.ts
