+++ news-aggregation/src/ui-kit/block/IMarkdownBlock.ts
