+++ news-aggregation/src/ui-kit/block/IPreviewBlock.ts
