+++ news-aggregation/src/ui-kit/block/ISectionBlock.ts
