+++ news-aggregation/src/ui-kit/element/IImageElement.ts
