+++ news-aggregation/src/ui-kit/element/IInteractionElement.ts
