+++ news-aggregation/src/ui-kit/element/IMultiStaticSelectElement.ts
