+++ news-aggregation/src/ui-kit/element/IPlainTextInputElement.ts
