+++ news-aggregation/src/ui-kit/element/IStaticSelectElement.ts
