+++ news-aggregation/src/utils/blocks.ts
