+++ news-aggregation/src/utils/sendNewsMessage.ts
